requirements.txt
streamlit==1.31.0
pandas==2.0.3
numpy==1.26.4
tensorflow==2.16.1
requests==2.31.0
pillow==10.3.0